src/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.1
beanie==1.11.12
black==22.10.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
distro==1.8.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.85.0
fastapi-users==10.1.5
fastapi-users-db-beanie==1.1.1
frozenlist==1.3.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
makefun==1.15.0
motor==3.0.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
yarl==1.8.1
